+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+Features introduced in the beta version stabilised and documented
 
 ## 3.0.0-beta.4 - 2024-06-21
 Better types compatibility for primitive types in `typedQuery`