+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 3.1.0-beta.0 - 2024-07-21
+
 ## 3.0.0 - 2024-06-21
 Features introduced in the beta version stabilised and documented